Converting class moved into separate shared library. Some methods converted back to synchronous because they don't use actually async calls.
#61
