+++ AdobeSdkService/AdobeSdkService.csproj
@@ -32,8 +32,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Adobe.DocumentCloud.Services" Version="1.0.0" />
-    <PackageReference Include="log4net" Version="2.0.8" />
+    <ProjectReference Include="..\HTMLToPDFConvertingDLL\HTMLToPDFConvertingDLL.csproj" />
   </ItemGroup>
