+++ HTML cleanup/HTMLCleanupDLL/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Adobe.DocumentCloud.Services" version="1.0.0" targetFramework="net461" />
   <package id="Common.Logging" version="3.4.1" targetFramework="net40-client" />
   <package id="Common.Logging.Core" version="3.4.1" targetFramework="net40-client" />
   <package id="itext7" version="7.1.6" targetFramework="net461" />
