+++ PdfCreator/PdfCreator.csproj
@@ -2,14 +2,13 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <AssemblyName>pdf_creator</AssemblyName>
     <StartupObject>PdfCreator.Program</StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
-     <PackageReference Include="Adobe.DocumentCloud.Services" Version="1.0.0" />
-     <PackageReference Include="log4net" Version="2.0.8" />
+    <ProjectReference Include="..\HTMLToPDFConvertingDLL\HTMLToPDFConvertingDLL.csproj" />
   </ItemGroup>
 
   <ItemGroup>
