+++ PdfService.sln
@@ -8,8 +8,16 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HTMLCleanupDLL", "HTML cleanup\HTMLCleanupDLL\HTMLCleanupDLL.csproj", "{F9BBFDAC-F625-4F46-8B47-9D99E1776264}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PdfCreator", "PdfCreator\PdfCreator.csproj", "{E57FC657-230C-4240-B1BC-ED3A8FFB7718}"
+	ProjectSection(ProjectDependencies) = postProject
+		{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508} = {4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}
+	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdobeSdkService", "AdobeSdkService\AdobeSdkService.csproj", "{E5929948-964A-4ED7-A308-F60DDFD0252C}"
+	ProjectSection(ProjectDependencies) = postProject
+		{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508} = {4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HTMLToPDFConvertingDLL", "HTMLToPDFConvertingDLL\HTMLToPDFConvertingDLL.csproj", "{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +41,10 @@ Global
 		{E5929948-964A-4ED7-A308-F60DDFD0252C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E5929948-964A-4ED7-A308-F60DDFD0252C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E5929948-964A-4ED7-A308-F60DDFD0252C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4BEE0C6B-EFE6-4F3D-8BD6-3D88A8F30508}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
